WebGen update.
git-svn-id: https://bl-toolkit.googlecode.com/svn/trunk/2@195 827d3e3c-cb35-0410-a708-6b02db8f8db3

+++ WebGen/Generator.cs
@@ -1,29 +1,40 @@
 using System;
 using System.IO;
+using System.Collections.Generic;
 
 using Rsdn.Framework.Formatting;
+using System.Globalization;
 
 namespace WebGen
 {
+	delegate FileAction FileActionHandler(string ext);
+
 	class Generator
 	{
-		string _sourcePath;
-		string _destFolder;
+		string            _sourcePath;
+		string            _destFolder;
+		FileActionHandler _fileAction;
 
-		public void Generate(string sourcePath, string destFolder)
+		public void Generate(
+			string templateFileName, string[] path,
+			string destFolder, string sourcePath,
+			bool cleanUp, bool createIndex, FileActionHandler fileAction)
 		{
 			_sourcePath = Path.GetFullPath(sourcePath);
 			_destFolder = Path.GetFullPath(destFolder);
+			_fileAction = fileAction;
 
-			CleanUp();
+			if (cleanUp)
+				CleanUp();
+
 			CreateDestFolder();
 
 			string template = null;
 
-			using (StreamReader sr = File.OpenText(Path.Combine(_sourcePath, "template.html")))
+			using (StreamReader sr = File.OpenText(templateFileName))
 				template = sr.ReadToEnd();
 
-			GenerateContent(template, new string[0]);
+			GenerateContent(template, path, createIndex);
 		}
 
 		private void CreateDestFolder()
@@ -40,7 +51,8 @@ namespace WebGen
 			Action<string> clean = null; clean = delegate(string path)
 			{
 				foreach (string file in Directory.GetFiles(path))
-					try { File.Delete(file); } catch {}
+					if (Path.GetExtension(file).ToLower() != ".zip")
+						try { File.Delete(file); } catch {}
 
 				foreach (string dir in Directory.GetDirectories(path))
 				{
@@ -52,7 +64,7 @@ namespace WebGen
 			clean(_destFolder);
 		}
 
-		private void GenerateContent(string template, string[] path)
+		private bool GenerateContent(string template, string[] path, bool createIndex)
 		{
 			string folder     = string.Join("/", path);
 			string destFolder = Path.Combine(_destFolder, folder);
@@ -61,42 +73,60 @@ namespace WebGen
 			for (int i = 0; i < path.Length; i++)
 				backPath += "../";
 
-			if (Directory.Exists(destFolder) == false)
-				Directory.CreateDirectory(destFolder);
-
 			string   sourcePath  = Path.Combine(_sourcePath, folder);
 			string[] sourceFiles = Directory.GetFiles(sourcePath);
-			string   backLinks   = GeneratePath(path, backPath);
+
+			List<string> files   = new List<string>();
+			List<string> folders = new List<string>();
 
 			foreach (string fileName in sourceFiles)
 			{
 				if (fileName.ToLower().EndsWith("template.html"))
 					continue;
 
+				string backLinks = GeneratePath(path, backPath, fileName);
+
 				string destName = Path.Combine(destFolder, Path.GetFileName(fileName));
 				string ext      = Path.GetExtension(destName).ToLower();
 
 				Console.WriteLine(destName);
 
-				switch (ext)
+				switch (_fileAction(fileName))
 				{
-					case ".config":
-					case ".xml":
-					case ".sql":
+					case FileAction.Skip:
 						break;
-					case ".htm":
-						using (StreamWriter sw = File.CreateText(destName))
-						using (StreamReader sr = File.OpenText(fileName))
-						{
-							string source = GenerateSource(sr.ReadToEnd());
 
-							sw.WriteLine(string.Format(template, source, backPath, backLinks));
+					case FileAction.Copy:
+						File.Copy(fileName, destName, true);
+						break;
+
+					case FileAction.Process:
+						if (Directory.Exists(destFolder) == false)
+							Directory.CreateDirectory(destFolder);
+
+						switch (ext)
+						{
+							case ".htm":
+							case ".html":
+								using (StreamWriter sw = File.CreateText(destName))
+								using (StreamReader sr = File.OpenText(fileName))
+								{
+									string source = GenerateSource(sr.ReadToEnd());
+									sw.WriteLine(string.Format(template, source, backPath, backLinks));
+								}
+								break;
+
+							case ".cs":
+								using (StreamWriter sw = File.CreateText(destName + ".htm"))
+								{
+									string source = GenerateSource("<% " + fileName + " %>");
+									sw.WriteLine(string.Format(template, source, backPath, backLinks));
+								}
+								break;
 						}
 
-						break;
+						files.Add(fileName);
 
-					default:
-						File.Copy(fileName, destName, true);
 						break;
 				}
 			}
@@ -116,8 +146,51 @@ namespace WebGen
 
 				newPath[path.Length] = dirName;
 
-				GenerateContent(template, newPath);
+				if (GenerateContent(template, newPath, createIndex))
+					folders.Add(dir);
 			}
+
+			if (files.Count > 0 || folders.Count > 0)
+			{
+				string indexName = destFolder + "/index.htm";
+
+				if (createIndex && File.Exists(indexName) == false)
+				{
+					string str = "";
+
+					folders.Sort();
+
+					foreach (string s in folders)
+						str += string.Format("&#8226; <a href='{0}/index.htm'>{0}</a><br>",
+							Path.GetFileName(s));
+
+					if (str.Length > 0)
+						str += "<br>";
+
+					files.Sort();
+
+					foreach (string s in files)
+						str += string.Format(
+							s.EndsWith(".htm",  true, CultureInfo.CurrentCulture) ||
+							s.EndsWith(".html", true, CultureInfo.CurrentCulture)?
+								"&#8226; <a href='{0}'>{0}</a><br>":
+								"&#8226; <a href='{0}.htm'>{0}</a><br>",
+							Path.GetFileName(s));
+
+					using (StreamWriter sw = File.CreateText(indexName))
+					{
+						sw.WriteLine(string.Format(
+							template,
+							str,
+							backPath,
+							GeneratePath(path, backPath, "@@@").Replace(".@@@", "")));
+					}
+				}
+
+				return true;
+			}
+
+			return false;
 		}
 
 		private string GenerateSource(string source)
@@ -127,7 +200,7 @@ namespace WebGen
 				 end = source.IndexOf("%>", idx + 2);
 				 idx >= 0 &&
 				 end >= 0;
-				 idx = source.IndexOf("<%", end + 2), 
+				 idx = source.IndexOf("<%", end + 2),
 				 end = source.IndexOf("%>", idx + 2))
 			{
 				string startSource = source.Substring(0, idx);
@@ -182,27 +255,53 @@ namespace WebGen
 			return source;
 		}
 
-		private string GeneratePath(string[] path, string backPath)
+		private string GeneratePath(string[] path, string backPath, string fileName)
 		{
+			if (path.Length == 0)
+				return "";
+
 			string backLinks = "";
 
-			if (path.Length > 1 && path[0] == "Doc")
+			switch (path[0])
 			{
-				backLinks += string.Format(
-					"<br><nobr>&nbsp;&nbsp;<small><a class='m' href='{0}index.htm'>BLToolkit</a>",
-					backPath);
+				case "Doc":
+					backLinks += string.Format(
+						"<br><nobr>&nbsp;&nbsp;<small><a class='m' href='{0}index.htm'>BLToolkit</a>",
+						backPath);
 
-				for (int i = 1; i < path.Length; i++)
-				{
-					string parent = "";
+					for (int i = 1; i < path.Length; i++)
+					{
+						string parent = "";
 
-					for (int j = i + 1; j < path.Length; j++)
-						parent += "../";
+						for (int j = i + 1; j < path.Length; j++)
+							parent += "../";
 
-					backLinks += string.Format(".<a class='m' href='{0}index.htm'>{1}</a>", parent, path[i]);
-				}
+						backLinks += string.Format(".<a class='m' href='{0}index.htm'>{1}</a>", parent, path[i]);
+					}
 
-				backLinks += "<small></nobr></br>";
+					backLinks += "<small></nobr></br>";
+
+					break;
+
+				case "Source":
+					backLinks += string.Format(
+						"<br><nobr>&nbsp;&nbsp;<small><a class='m' href='{0}/Source/index.htm'>BLToolkit</a>",
+						backPath);
+
+					for (int i = 1; i < path.Length; i++)
+					{
+						string parent = "";
+
+						for (int j = i + 1; j < path.Length; j++)
+							parent += "../";
+
+						backLinks += string.Format(".<a class='m' href='{0}index.htm'>{1}</a>", parent, path[i]);
+					}
+
+					backLinks += "." + Path.GetFileNameWithoutExtension(fileName);
+					backLinks += "<small></nobr></br>";
+
+					break;
 			}
 
 			return backLinks;
